+++ osmd-universal-viewer/src/nwc2xml-lib/index.d.ts
@@ -1,0 +1,1 @@
+export function convertNWCToMusicXML(buffer: Uint8Array | ArrayBuffer): string;